feat: Optimize & split components
