+++ src/renderer/src/components/session-detail/ai-panel/index.ts
@@ -1,0 +1,1 @@
+export { MessagePanel } from './MessagePanel'
