+++ src/renderer/src/components/session-detail/element-inspector/index.ts
@@ -1,0 +1,2 @@
+export { ElementInspectorPanel } from './ElementInspectorPanel'
+export type { ElementEditDraft } from './ElementInspectorPanel'
