+++ src/renderer/src/components/session-detail/preview/index.ts
@@ -1,0 +1,1 @@
+export { PreviewStage } from './PreviewStage'
