+++ src/renderer/src/components/session-detail/shared/index.ts
@@ -1,0 +1,7 @@
+export {
+  buildImageMessageCacheKey,
+  imageHistoryToMessages,
+  mergeImageMessages
+} from './imageMessageUtils'
+export { normalizePagesForSelection } from './pageUtils'
+export type { ChatType, SessionPreviewPage } from './types'
