+++ src/renderer/src/components/session-detail/shared/types.ts
