+++ src/renderer/src/components/session-detail/sidebar/index.ts
@@ -1,0 +1,1 @@
+export { PageSidebar } from './PageSidebar'
