+++ src/renderer/src/components/session-detail/speech/index.ts
@@ -1,0 +1,1 @@
+export { SpeechScriptDrawer } from './SpeechScriptDrawer'
