+++ src/renderer/src/components/session-detail/toolbar/SessionToolbar.tsx
@@ -14,56 +14,38 @@ import {
 } from 'lucide-react'
 import { cn } from '@renderer/lib/utils'
 import { useSessionDetailUiStore } from '@renderer/store'
-import { Button } from '../ui/Button'
+import { Button } from '../../ui/Button'
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuTrigger
-} from '../ui/DropdownMenu'
-import { Tooltip, TooltipContent, TooltipTrigger } from '../ui/Tooltip'
+} from '../../ui/DropdownMenu'
+import { Tooltip, TooltipContent, TooltipTrigger } from '../../ui/Tooltip'
 import { useT } from '@renderer/i18n'
+import { SaveTemplateDialog } from '../../templates/SaveTemplateDialog'
+import { useSessionToolbarController } from './useSessionToolbarController'
 
 const toolbarButtonClass =
   'h-7 rounded-[8px] border-transparent bg-[#e8e0d0]/72 px-2.5 text-[11px] text-[#3e4a32] shadow-[0_4px_10px_rgba(86,72,53,0.08)] hover:bg-[#d4e4c1]/78'
 const toolbarIconClass = 'mr-1.5 h-3.5 w-3.5'
 const dropdownItemIconClass = 'mr-2 h-3.5 w-3.5 text-[#6b7280]'
 
-export function SessionToolbar({
-  hasPages,
-  historyDisabled = false,
-  canPreview,
-  canRevealFile,
-  onExportPdf,
-  onExportPng,
-  onExportPptx,
-  onExportSessionZip,
-  onExportSlidePack,
-  onOpenHistory,
-  onOpenPreview,
-  onRevealFile,
-  onSaveTemplate,
-  onPresent
-}: {
-  hasPages: boolean
-  historyDisabled?: boolean
-  canPreview: boolean
-  canRevealFile: boolean
-  onExportPdf: () => void
-  onExportPng: () => void
-  onExportPptx: (options?: {
-    imageOnly?: boolean
-    embedFonts?: boolean | 'auto' | 'always' | 'never'
-  }) => void
-  onExportSessionZip: () => void
-  onExportSlidePack: () => void
-  onOpenHistory: () => void
-  onOpenPreview: () => void
-  onRevealFile: () => void
-  onSaveTemplate?: () => void
-  onPresent?: () => void
-}): React.JSX.Element {
+export function SessionToolbar({ sessionId }: { sessionId: string }): React.JSX.Element {
   const t = useT()
+  const {
+    hasPages,
+    historyDisabled,
+    canPreview,
+    canRevealFile,
+    saveTemplateOpen,
+    savingTemplate,
+    defaultTemplateName,
+    setSaveTemplateOpen,
+    handleSaveTemplate,
+    exportActions,
+    openHistory
+  } = useSessionToolbarController(sessionId)
 
   const isExportingPdf = useSessionDetailUiStore((state) => state.isExportingPdf)
   const isExportingPng = useSessionDetailUiStore((state) => state.isExportingPng)
@@ -89,7 +71,7 @@ export function SessionToolbar({
               variant="outline"
               size="sm"
               className={toolbarButtonClass}
-              onClick={onOpenHistory}
+              onClick={openHistory}
               disabled={historyDisabled || isExporting}
             >
               <History className={toolbarIconClass} />
@@ -101,7 +83,7 @@ export function SessionToolbar({
           </TooltipContent>
         </Tooltip>
       )}
-      {hasPages && onSaveTemplate && (
+      {hasPages && (
         <Tooltip>
           <TooltipTrigger asChild>
             <Button
@@ -109,7 +91,7 @@ export function SessionToolbar({
               variant="outline"
               size="sm"
               className={toolbarButtonClass}
-              onClick={onSaveTemplate}
+              onClick={() => setSaveTemplateOpen(true)}
               disabled={isExporting}
             >
               <LayoutTemplate className={toolbarIconClass} />
@@ -141,11 +123,11 @@ export function SessionToolbar({
             </Button>
           </DropdownMenuTrigger>
           <DropdownMenuContent align="start" className="min-w-[14rem]">
-            <DropdownMenuItem onClick={() => onExportPptx()}>
+            <DropdownMenuItem onClick={() => void exportActions.exportPptx()}>
               <Presentation className={dropdownItemIconClass} />
               {t('sessionDetail.exportPptxEditable')}
             </DropdownMenuItem>
-            <DropdownMenuItem onClick={() => onExportPptx({ imageOnly: true })}>
+            <DropdownMenuItem onClick={() => void exportActions.exportPptx({ imageOnly: true })}>
               <ImageIcon className={dropdownItemIconClass} />
               {t('sessionDetail.exportPptxImageOnly')}
             </DropdownMenuItem>
@@ -177,7 +159,10 @@ export function SessionToolbar({
             <TooltipContent>{t('sessionDetail.exportPackageTooltip')}</TooltipContent>
           </Tooltip>
           <DropdownMenuContent align="start" className="w-72">
-            <DropdownMenuItem className="items-start" onClick={onExportSlidePack}>
+            <DropdownMenuItem
+              className="items-start"
+              onClick={() => void exportActions.exportSlidePack()}
+            >
               <Package className={cn(dropdownItemIconClass, 'mt-0.5')} />
               <span className="flex min-w-0 flex-1 flex-col gap-0.5 whitespace-normal">
                 <span>{t('sessionDetail.exportSlidePack')}</span>
@@ -186,7 +171,10 @@ export function SessionToolbar({
                 </span>
               </span>
             </DropdownMenuItem>
-            <DropdownMenuItem className="items-start" onClick={onExportSessionZip}>
+            <DropdownMenuItem
+              className="items-start"
+              onClick={() => void exportActions.exportSessionZip()}
+            >
               <Archive className={cn(dropdownItemIconClass, 'mt-0.5')} />
               <span className="flex min-w-0 flex-1 flex-col gap-0.5 whitespace-normal">
                 <span>{t('sessionDetail.exportSessionZip')}</span>
@@ -218,11 +206,11 @@ export function SessionToolbar({
             </Button>
           </DropdownMenuTrigger>
           <DropdownMenuContent align="start" className="min-w-[12rem]">
-            <DropdownMenuItem onClick={onExportPng}>
+            <DropdownMenuItem onClick={() => void exportActions.exportPng()}>
               <ImageIcon className={dropdownItemIconClass} />
               {t('sessionDetail.exportPng')}
             </DropdownMenuItem>
-            <DropdownMenuItem onClick={onExportPdf}>
+            <DropdownMenuItem onClick={() => void exportActions.exportPdf()}>
               <FileDown className={dropdownItemIconClass} />
               {t('sessionDetail.exportPdf')}
             </DropdownMenuItem>
@@ -237,7 +225,7 @@ export function SessionToolbar({
               variant="outline"
               size="sm"
               className={toolbarButtonClass}
-              onClick={onOpenPreview}
+              onClick={() => void exportActions.openProjectPreview()}
             >
               <ExternalLink className={toolbarIconClass} />
               {t('sessionDetail.preview')}
@@ -246,7 +234,7 @@ export function SessionToolbar({
           <TooltipContent>{t('sessionDetail.previewTooltip')}</TooltipContent>
         </Tooltip>
       )}
-      {hasPages && onPresent && (
+      {hasPages && (
         <Tooltip>
           <TooltipTrigger asChild>
             <Button
@@ -254,7 +242,7 @@ export function SessionToolbar({
               variant="outline"
               size="sm"
               className={toolbarButtonClass}
-              onClick={onPresent}
+              onClick={() => void exportActions.openPresentation()}
             >
               <Monitor className={toolbarIconClass} />
               {t('sessionDetail.present')}
@@ -269,12 +257,19 @@ export function SessionToolbar({
           variant="outline"
           size="sm"
           className={toolbarButtonClass}
-          onClick={onRevealFile}
+          onClick={() => void exportActions.revealSelectedPageFile()}
         >
           <FileSearch className={toolbarIconClass} />
           {t('sessionDetail.revealFile')}
         </Button>
       )}
+      <SaveTemplateDialog
+        open={saveTemplateOpen}
+        defaultName={defaultTemplateName}
+        saving={savingTemplate}
+        onOpenChange={setSaveTemplateOpen}
+        onSubmit={(payload) => void handleSaveTemplate(payload)}
+      />
     </>
   )
 }
