+++ src/renderer/src/components/session-detail/toolbar/index.ts
@@ -1,0 +1,1 @@
+export { SessionToolbar } from './SessionToolbar'
