+++ src/renderer/src/components/session-detail/workspace/WorkspaceRibbon.tsx
