+++ src/renderer/src/components/session-detail/workspace/index.ts
@@ -1,0 +1,2 @@
+export { WorkspaceRibbon } from './WorkspaceRibbon'
+export { EmptyEditWorkbenchPanel } from './workbench'
