+++ src/renderer/src/components/session-detail/workspace/toolbar/DynamicToolRow.tsx
