+++ src/renderer/src/components/session-detail/workspace/toolbar/WorkspaceTabs.tsx
