+++ src/renderer/src/components/session-detail/workspace/toolbar/tool-rows/ToolRowShell.tsx
