+++ src/renderer/src/components/session-detail/workspace/toolbar/types.ts
