+++ src/renderer/src/components/session-detail/workspace/workbench/WorkbenchPanelShell.tsx
