+++ src/renderer/src/components/session-detail/workspace/workbench/index.ts
