+++ src/renderer/src/components/session-detail/workspace/workbench/styles.ts
